+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [BREAKING]: Rename `Algorithm` to `Interval` for easier to understand.
+- [BREAKING]: Rename `run_failable` to `run_fallible`.